undo element resolver class
